Rework of store. split region detail drawer state from mapStore
